Organizes tests for trees into their own directory structure.
